Added start and end dates for courses and have course enrollment inherit that by default
